Move util function to utils.py
